+++ peer/peerlistv2/doc.go
@@ -18,11 +18,6 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-// Package peerlist is depracted in favor of
-// `go.uber.org/yarpc/peer/peerlistv2` which can additionally convey peer list
-// identifiers to the peerlist.Implementation without a wrapper type, allowing
-// a peer list updater to communicate shard information for example.
-//
 // Package peerlist provides a utility for managing peer availability with a
 // separate implementation of peer selection from just among available peers.
 // The peer list implements the peer.ChooserList interface and accepts a
@@ -31,7 +26,7 @@
 //
 // The example is an implementation of peer.ChooserList using a random peer selection
 // strategy, returned by newRandomListImplementation(), implementing
-// peer.ListImplementation.
+// peerlist.Implementation.
 //
 //   type List struct {
 //   	*peerlist.List